Resolve build warnings
