+++ bench/Main.hs
@@ -12,8 +12,8 @@ import           Data.Text            (Text)
 import qualified Data.Text            as T
 import           Data.Time
 -------------------------------------------------------------------------------
+import           System.Cron
 import           System.Cron.Parser
-import           System.Cron2
 -------------------------------------------------------------------------------
